Refactor adapter CLI into package
