+++ projects/04-llm-adapter/adapter/cli/__main__.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+from . import main
+
+if __name__ == "__main__":  # pragma: no cover
+    from sys import exit as _exit
+
+    _exit(main())
